Moved Sim and utils out of envs folder
